requirements.txt
anyio==4.4.0
asttokens==2.4.1
black==24.4.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docutils==0.21.2
executing==2.0.1
fonttools==4.53.0
h11==0.14.0
idna==3.7
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
kiwisolver==1.4.5
marimo==0.6.26
Markdown==3.6
matplotlib==3.9.0
matplotlib-inline==0.1.7
mypy-extensions==1.0.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure-eval==0.2.2
Pygments==2.18.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.37.2
tomlkit==0.12.5
traitlets==5.14.3
tzdata==2024.1
uvicorn==0.30.1
wcwidth==0.2.13
websockets==12.0
